ConstantPool object that interns int/float/string for unique IR::Constant Value since instructions only use constants as direct operands. Different from MLIR literal = instuction owned by basic block

+++ src/core/frontend/ast/ast_node.hpp
@@ -157,9 +157,9 @@ namespace Syntax
 
     struct FloatLiteralExpr
     {
-        long double value_;
+        double value_;
 
-        FloatLiteralExpr(long double value) noexcept :
+        FloatLiteralExpr(double value) noexcept :
             value_{ value } {}
     };
