+++ src/core/frontend/sema/sema_node.hpp
@@ -145,9 +145,9 @@ namespace Sema
 
     struct FloatLiteralExpr
     {
-        long double value_;
+        double value_;
 
-        FloatLiteralExpr(long double value) noexcept :
+        FloatLiteralExpr(double value) noexcept :
             value_{ value } {}
     };
